chore : 프로젝트 구조 수정 - domain 경로 제거
