+++ src/main/java/com/kosa/fillinv/category/repository/CategoryRepository.java
@@ -1,0 +1,9 @@
+package com.kosa.fillinv.category.repository;
+
+import com.kosa.fillinv.category.entity.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+}
