--- src/main/java/com/kosa/fillinv/domain/category/entity/Category.java
@@ -1,20 +1,0 @@
-package com.kosa.fillinv.domain.category.entity;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "categories")
-public class Category {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "category_id", nullable = false)
-    private Long id;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @ManyToOne
-    @JoinColumn(name = "parent_category_id")
-    private Category parentCategory;
-}
