--- src/main/java/com/kosa/fillinv/domain/category/repository/CategoryRepository.java
@@ -1,9 +1,0 @@
-package com.kosa.fillinv.domain.category.repository;
-
-import com.kosa.fillinv.domain.category.entity.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
