--- src/main/java/com/kosa/fillinv/domain/lesson/entity/AvailableTime.java
@@ -1,41 +1,0 @@
-package com.kosa.fillinv.domain.lesson.entity;
-
-import jakarta.persistence.*;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-
-@Entity
-@Table(name = "available_times")
-public class AvailableTime {
-
-    @Id
-    @Column(name = "available_time_id", nullable = false)
-    private String id;
-
-    @Column(name = "date", nullable = false)
-    private LocalDate date;
-
-    @Column(name = "start_time", nullable = false)
-    private LocalTime startTime;
-
-    @Column(name = "end_time", nullable = false)
-    private LocalTime endTime;
-
-    @Column(name = "price", nullable = false)
-    private Integer price;
-
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
-    @Column(name = "deleted_at")
-    private LocalDateTime deletedAt;
-
-    @ManyToOne
-    @JoinColumn(name = "lesson_id", nullable = false)
-    private Lesson lesson;
-}
