--- src/main/java/com/kosa/fillinv/domain/lesson/entity/LessonType.java
@@ -1,7 +1,0 @@
-package com.kosa.fillinv.domain.lesson.entity;
-
-public enum LessonType {
-    MENTORING,
-    ONEDAY,
-    STUDY
-}
