--- src/main/java/com/kosa/fillinv/domain/member/entity/Member.java
@@ -1,35 +1,0 @@
-package com.kosa.fillinv.domain.member.entity;
-
-import jakarta.persistence.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "members")
-public class Member {
-
-    @Id
-    @Column(name = "member_id", nullable = false)
-    private String id;
-
-    @Column(name = "nickname", nullable = false)
-    private String nickname;
-
-    @Column(name = "phone_num", nullable = false)
-    private String phoneNum;
-
-    @Column(name = "email", nullable = false)
-    private String email;
-
-    @Column(name = "password", nullable = false)
-    private String password;
-
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
-    @Column(name = "deleted_at")
-    private LocalDateTime deletedAt;
-}
