--- src/main/java/com/kosa/fillinv/domain/member/entity/Profile.java
@@ -1,36 +1,0 @@
-package com.kosa.fillinv.domain.member.entity;
-
-import jakarta.persistence.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "profiles")
-public class Profile {
-
-    @Id
-    @Column(name = "member_id", nullable = false)
-    private String memberId;
-
-    @OneToOne
-    @MapsId
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    private String image;
-
-    @Column(name = "introduce", nullable = false)
-    private String introduce;
-
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
-    @Column(name = "deleted_at")
-    private LocalDateTime deletedAt;
-
-    @Column(name = "category_id", nullable = false)
-    private Long categoryId;
-}
