--- src/main/java/com/kosa/fillinv/domain/member/repository/ProfileRepository.java
@@ -1,9 +1,0 @@
-package com.kosa.fillinv.domain.member.repository;
-
-import com.kosa.fillinv.domain.member.entity.Profile;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProfileRepository extends JpaRepository<Profile, String> {
-}
