--- src/main/java/com/kosa/fillinv/domain/payment/entity/PaymentHistoryType.java
@@ -1,7 +1,0 @@
-package com.kosa.fillinv.domain.payment.entity;
-
-public enum PaymentHistoryType {
-    PAYMENT,
-    FULL_REFUND,
-    PARTIAL_REFUND
-}
