--- src/main/java/com/kosa/fillinv/domain/payment/repository/PaymentHistoryRepository.java
@@ -1,9 +1,0 @@
-package com.kosa.fillinv.domain.payment.repository;
-
-import com.kosa.fillinv.domain.payment.entity.PaymentHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PaymentHistoryRepository extends JpaRepository<PaymentHistory, String> {
-}
