--- src/main/java/com/kosa/fillinv/domain/review/repository/ReviewRepository.java
@@ -1,9 +1,0 @@
-package com.kosa.fillinv.domain.review.repository;
-
-import com.kosa.fillinv.domain.review.entity.Review;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ReviewRepository extends JpaRepository<Review, String> {
-}
