--- src/main/java/com/kosa/fillinv/domain/schedule/repository/ScheduleRepository.java
@@ -1,9 +1,0 @@
-package com.kosa.fillinv.domain.schedule.repository;
-
-import com.kosa.fillinv.domain.schedule.entity.Schedule;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ScheduleRepository extends JpaRepository<Schedule, String> {
-}
