+++ src/main/java/com/kosa/fillinv/lesson/entity/LessonType.java
@@ -1,0 +1,7 @@
+package com.kosa.fillinv.lesson.entity;
+
+public enum LessonType {
+    MENTORING,
+    ONEDAY,
+    STUDY
+}
