+++ src/main/java/com/kosa/fillinv/lesson/repository/LessonRepository.java
@@ -1,0 +1,9 @@
+package com.kosa.fillinv.lesson.repository;
+
+import com.kosa.fillinv.lesson.entity.Lesson;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface LessonRepository extends JpaRepository<Lesson, String> {
+}
