+++ src/main/java/com/kosa/fillinv/member/repository/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.kosa.fillinv.member.repository;
+
+import com.kosa.fillinv.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, String> {
+}
