+++ src/main/java/com/kosa/fillinv/member/repository/ProfileRepository.java
@@ -1,0 +1,9 @@
+package com.kosa.fillinv.member.repository;
+
+import com.kosa.fillinv.member.entity.Profile;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProfileRepository extends JpaRepository<Profile, String> {
+}
