+++ src/main/java/com/kosa/fillinv/payment/entity/PaymentHistoryType.java
@@ -1,0 +1,7 @@
+package com.kosa.fillinv.payment.entity;
+
+public enum PaymentHistoryType {
+    PAYMENT,
+    FULL_REFUND,
+    PARTIAL_REFUND
+}
