+++ src/main/java/com/kosa/fillinv/payment/repository/PaymentHistoryRepository.java
@@ -1,0 +1,9 @@
+package com.kosa.fillinv.payment.repository;
+
+import com.kosa.fillinv.payment.entity.PaymentHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PaymentHistoryRepository extends JpaRepository<PaymentHistory, String> {
+}
