+++ src/main/java/com/kosa/fillinv/review/repository/ReviewRepository.java
@@ -1,0 +1,9 @@
+package com.kosa.fillinv.review.repository;
+
+import com.kosa.fillinv.review.entity.Review;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ReviewRepository extends JpaRepository<Review, String> {
+}
