+++ src/main/java/com/kosa/fillinv/schedule/entity/Schedule.java
@@ -1,0 +1,73 @@
+package com.kosa.fillinv.schedule.entity;
+
+import jakarta.persistence.*;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+
+@Entity
+@Table(name = "schedules")
+public class Schedule {
+
+    @Id
+    @Column(name = "schedule_id", nullable = false)
+    private String id;
+
+    @Column(name = "date", nullable = false)
+    private LocalDate date;
+
+    @Column(name = "start_time", nullable = false)
+    private LocalTime startTime;
+
+    @Column(name = "end_time", nullable = false)
+    private LocalTime endTime;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "status", nullable = false)
+    private ScheduleStatus status;
+
+    @Column(name = "request_content")
+    private String requestContent;
+
+    /* ===== Lesson Snapshot ===== */
+    @Column(name = "lesson_type", nullable = false)
+    private String lessonType;
+
+    @Column(name = "lesson_description", nullable = false)
+    private String lessonDescription;
+
+    @Column(name = "lesson_location", nullable = false)
+    private String lessonLocation;
+
+    @Column(name = "lesson_category_name", nullable = false)
+    private String lessonCategoryName;
+
+    /* ===== Option Snapshot ===== */
+    @Column(name = "option_name", nullable = false)
+    private String optionName;
+
+    @Column(name = "option_minute", nullable = false)
+    private Integer optionMinute;
+
+    @Column(name = "price", nullable = false)
+    private Integer price;
+
+    @Column(name = "created_at")
+    private LocalDateTime createdAt;
+
+    @Column(name = "updated_at")
+    private LocalDateTime updatedAt;
+
+    @Column(name = "deleted_at")
+    private LocalDateTime deletedAt;
+
+    @Column(name = "lesson_id")
+    private String lessonId;
+
+    @Column(name = "mentee_id")
+    private String mentee;
+
+    @Column(name = "option_id")
+    private String optionId;
+}
