+++ src/main/java/com/kosa/fillinv/schedule/entity/ScheduleStatus.java
@@ -1,0 +1,19 @@
+package com.kosa.fillinv.schedule.entity;
+
+public enum ScheduleStatus {
+    PAYMENT_PENDING("결제 대기"),
+    APPROVAL_PENDING("승인 대기"),
+    APPROVED("승인"),
+    CANCELED("취소"),
+    COMPLETED("완료");
+
+    private final String description;
+
+    ScheduleStatus(String description) {
+        this.description = description;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+}
