+++ src/main/java/com/kosa/fillinv/schedule/repository/ScheduleRepository.java
@@ -1,0 +1,9 @@
+package com.kosa.fillinv.schedule.repository;
+
+import com.kosa.fillinv.schedule.entity.Schedule;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ScheduleRepository extends JpaRepository<Schedule, String> {
+}
